Address Node test review feedback
