+++ meshtastic/tests/test_node_channels.py
@@ -22,6 +22,7 @@ from ..serial_interface import SerialInterface
 from ..util import fromPSK
 
 from ._node_legacy_support import (
+    _decode_channel_set_from_url,
     _get_mock_call_arg,
     _make_fake_send_admin,
 )
@@ -318,30 +319,70 @@ def test_setOwner_rejects_empty_or_whitespace_names(
 
 @pytest.mark.unit
 @pytest.mark.parametrize(
-    ("owner_kwargs", "expected_patterns"),
+    ("owner_kwargs", "expected_patterns", "unexpected_patterns", "expected_short_name"),
     [
-        (
+        pytest.param(
             {"long_name": "ValidName", "short_name": "VN"},
             (
                 r"p\.set_owner\.long_name_set:True",
                 r"p\.set_owner\.short_name_set:True",
             ),
+            (
+                r"p\.set_owner\.is_licensed:True",
+                r"p\.set_owner\.is_unmessagable:True",
+            ),
+            "VN",
+            id="long-and-short",
         ),
-        (
+        pytest.param(
             {"short_name": "TST"},
             (r"p\.set_owner\.short_name_set:True",),
+            (
+                r"p\.set_owner\.long_name_set:True",
+                r"p\.set_owner\.is_licensed:True",
+                r"p\.set_owner\.is_unmessagable:True",
+            ),
+            "TST",
+            id="short-only",
         ),
-        (
+        pytest.param(
             {"long_name": "TestUser", "short_name": "TOOLONG"},
-            (r"p\.set_owner\.short_name_set:True",),
+            (
+                r"p\.set_owner\.long_name_set:True",
+                r"p\.set_owner\.short_name_set:True",
+            ),
+            (
+                r"p\.set_owner\.is_licensed:True",
+                r"p\.set_owner\.is_unmessagable:True",
+            ),
+            "TOOL",
+            id="short-name-truncated",
         ),
-        (
+        pytest.param(
             {"long_name": "LicensedUser", "is_licensed": True},
-            (r"p\.set_owner\.is_licensed:True",),
+            (
+                r"p\.set_owner\.long_name_set:True",
+                r"p\.set_owner\.is_licensed:True",
+            ),
+            (
+                r"p\.set_owner\.short_name_set:True",
+                r"p\.set_owner\.is_unmessagable:True",
+            ),
+            "",
+            id="licensed",
         ),
-        (
+        pytest.param(
             {"long_name": "TestUser", "is_unmessagable": True},
-            (r"p\.set_owner\.is_unmessagable:True",),
+            (
+                r"p\.set_owner\.long_name_set:True",
+                r"p\.set_owner\.is_unmessagable:True",
+            ),
+            (
+                r"p\.set_owner\.short_name_set:True",
+                r"p\.set_owner\.is_licensed:True",
+            ),
+            "",
+            id="unmessagable",
         ),
     ],
 )
@@ -350,16 +391,23 @@ def test_setOwner_logs_expected_fields_for_variants(
     autospec_local_node_iface: Callable[[type[Any]], MagicMock],
     owner_kwargs: dict[str, Any],
     expected_patterns: tuple[str, ...],
+    unexpected_patterns: tuple[str, ...],
+    expected_short_name: str,
 ) -> None:
-    """Test setOwner variants log the expected fields."""
+    """Test setOwner variants set only the requested fields and truncate short names."""
     iface = autospec_local_node_iface(MeshInterface)
     anode = Node(iface, 123, noProto=True)
+    anode._send_admin = MagicMock(return_value=mesh_pb2.MeshPacket())  # type: ignore[method-assign]
 
     with caplog.at_level(logging.DEBUG):
         anode.setOwner(**owner_kwargs)
 
     for pattern in expected_patterns:
         assert re.search(pattern, caplog.text, re.MULTILINE)
+    for pattern in unexpected_patterns:
+        assert not re.search(pattern, caplog.text, re.MULTILINE)
+    sent_msg = anode._send_admin.call_args.args[0]
+    assert sent_msg.set_owner.short_name == expected_short_name
 
 
 @pytest.mark.unit
@@ -374,7 +422,11 @@ def test_waitForConfig_timeout(
     anode._timeout.waitForSet.return_value = False
 
     result = anode.waitForConfig()
+
     assert result is False
+    wait_call = anode._timeout.waitForSet.call_args
+    assert wait_call.kwargs["attrs"] == ("is_set",)
+    assert getattr(wait_call.args[0], "_node", None) is anode
 
 
 @pytest.mark.unit
@@ -392,8 +444,12 @@ def test_waitForConfig_success(
     anode._timeout = MagicMock()
     anode._timeout.waitForSet.return_value = True
 
-    result = anode.waitForConfig()
+    result = anode.waitForConfig(attribute="lora")
+
     assert result is True
+    wait_call = anode._timeout.waitForSet.call_args
+    assert wait_call.kwargs["attrs"] == ("is_set",)
+    assert getattr(wait_call.args[0], "_name", None) == "lora"
 
 
 @pytest.mark.unit
@@ -479,6 +535,7 @@ def test_requestConfig_with_module_config_descriptor(
     assert len(sent_messages) == 1
     sent_msg = sent_messages[0]
     # mqtt field has index 0, should be set as get_module_config_request
+    assert sent_msg.WhichOneof("payload_variant") == "get_module_config_request"
     assert sent_msg.get_module_config_request == 0
 
 
@@ -635,6 +692,7 @@ def test_deleteChannel_switches_admin_index_after_rewriting_former_admin_slot(
         for call in anode._send_admin.call_args_list
     ]
     # Keep old admin index (2) through rewrite of old slot 2, then switch to 1.
+    assert len(admin_indexes) > 2
     assert admin_indexes[:2] == [2, 2]
     assert all(index == 1 for index in admin_indexes[2:])
 
@@ -869,6 +927,10 @@ def test_deleteChannel_rewrite_uses_snapshot_when_channels_change_after_lock_rel
     # channel snapshot list.
     assert anode.channels is None
     assert anode._send_admin.call_count == CHANNEL_LIMIT
+    written_indexes = [
+        call.args[0].set_channel.index for call in anode._send_admin.call_args_list
+    ]
+    assert written_indexes == list(range(CHANNEL_LIMIT))
 
 
 @pytest.mark.unit
@@ -882,3 +944,164 @@ def test_channel_lookup_helpers_return_none_when_no_match(
     assert anode.getChannelByName("missing") is None
     assert anode.getDisabledChannel() is None
     assert anode._get_admin_channel_index() == 0
+
+@pytest.mark.unit
+def test_fixup_channels_truncates_and_reindexes_to_limit(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """_fixup_channels should truncate over-limit input and maintain contiguous indices."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    anode.channels = [
+        Channel(index=i, role=Channel.Role.SECONDARY) for i in range(CHANNEL_LIMIT + 2)
+    ]
+
+    anode._fixup_channels()
+
+    assert anode.channels is not None
+    assert len(anode.channels) == CHANNEL_LIMIT
+    assert [ch.index for ch in anode.channels] == list(range(CHANNEL_LIMIT))
+
+@pytest.mark.unit
+def test_fill_channels_handles_none_and_pads_to_limit(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """_fill_channels should no-op for None and pad existing channel lists to max size."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    anode.channels = None
+    anode._fill_channels()
+    assert anode.channels is None
+
+    anode.channels = [Channel(index=0, role=Channel.Role.PRIMARY)]
+    anode._fill_channels()
+    assert anode.channels is not None
+    assert len(anode.channels) == CHANNEL_LIMIT
+    assert anode.channels[-1].role == Channel.Role.DISABLED
+
+@pytest.mark.unit
+def test_onResponseRequestChannel_routing_paths(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """Routing failures should terminate the request; success should await ADMIN_APP."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    anode._request_channel = MagicMock()  # type: ignore[method-assign]
+
+    anode.onResponseRequestChannel(
+        {
+            "decoded": {
+                "portnum": "ROUTING_APP",
+                "routing": {"errorReason": "NO_ROUTE"},
+            }
+        }
+    )
+    assert anode._channel_response_runtime.has_channel_request_failed() is True
+    anode._request_channel.assert_not_called()
+
+    channel = Channel(index=3, role=Channel.Role.SECONDARY)
+    anode.partialChannels = [channel]
+    anode._channel_response_runtime.mark_channel_request_sent(3)
+    anode.onResponseRequestChannel(
+        {"decoded": {"portnum": "ROUTING_APP", "routing": {"errorReason": "NONE"}}}
+    )
+
+    assert anode._channel_response_runtime.has_channel_request_failed() is False
+    assert anode.partialChannels == [channel]
+    anode._request_channel.assert_not_called()
+
+@pytest.mark.unit
+def test_onResponseRequestChannel_handles_partial_and_final_channel(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """OnResponseRequestChannel should request next channel until the final channel arrives."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    anode._request_channel = MagicMock()  # type: ignore[method-assign]
+
+    partial = Channel(index=2, role=Channel.Role.SECONDARY)
+    anode._channel_response_runtime.mark_channel_request_sent(2)
+    anode.onResponseRequestChannel(
+        {
+            "decoded": {
+                "portnum": "ADMIN_APP",
+                "admin": {"raw": MagicMock(get_channel_response=partial)},
+            }
+        }
+    )
+    anode._request_channel.assert_called_once_with(3)
+
+    final = Channel(index=CHANNEL_LIMIT - 1, role=Channel.Role.SECONDARY)
+    anode._request_channel.reset_mock()
+    anode._channel_response_runtime.mark_channel_request_sent(CHANNEL_LIMIT - 1)
+    anode.onResponseRequestChannel(
+        {
+            "decoded": {
+                "portnum": "ADMIN_APP",
+                "admin": {"raw": MagicMock(get_channel_response=final)},
+            }
+        }
+    )
+    anode._request_channel.assert_not_called()
+    assert anode.channels is not None
+    assert len(anode.channels) == CHANNEL_LIMIT
+
+@pytest.mark.unit
+def test_get_channels_with_hash_handles_missing_fields(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """_get_channels_with_hash should emit hashes only when both name and PSK are present."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    with_hash = Channel(index=0, role=Channel.Role.PRIMARY)
+    with_hash.settings.name = "hash-me"
+    with_hash.settings.psk = b"\x01\x02"
+    without_hash = Channel(index=1, role=Channel.Role.SECONDARY)
+    anode.channels = [with_hash, without_hash]
+
+    entries = anode._get_channels_with_hash()
+
+    assert len(entries) == 2
+    assert entries[0]["hash"] is not None
+    assert entries[1]["hash"] is None
+
+@pytest.mark.unit
+def test_fixup_channels_returns_immediately_when_channels_none(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """_fixup_channels should no-op when channels are unset."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    anode.channels = None
+
+    anode._fixup_channels()
+
+    assert anode.channels is None
+
+@pytest.mark.unit
+def test_getURL_requests_lora_when_local_config_empty(
+    autospec_local_node_iface: Callable[[type[Any]], MagicMock],
+) -> None:
+    """GetURL should request lora config when localConfig has no populated fields."""
+    anode = Node(autospec_local_node_iface(MeshInterface), "!12345678", noProto=True)
+    primary = Channel(index=0, role=Channel.Role.PRIMARY)
+    primary.settings.name = "primary"
+    primary.settings.psk = b"\x01"
+    secondary = Channel(index=1, role=Channel.Role.SECONDARY)
+    secondary.settings.name = "secondary"
+    secondary.settings.psk = b"\x02"
+    disabled = Channel(index=2, role=Channel.Role.DISABLED)
+    anode.channels = [primary, secondary, disabled]
+    anode.requestConfig = MagicMock()  # type: ignore[method-assign]
+
+    def _populate_lora_and_return_true(*, attribute: str = "channels") -> bool:
+        _ = attribute
+        anode.localConfig.lora.hop_limit = 3
+        return True
+
+    anode.waitForConfig = MagicMock(  # type: ignore[method-assign]
+        side_effect=_populate_lora_and_return_true
+    )
+
+    url = anode.getURL(includeAll=False)
+
+    anode.requestConfig.assert_called_once_with(
+        anode.localConfig.DESCRIPTOR.fields_by_name["lora"]
+    )
+    channel_set = _decode_channel_set_from_url(url)
+    assert len(channel_set.settings) == 1
+    assert channel_set.settings[0].name == "primary"
